move image_log.py to utils.py
