+++ mapillary_tools/geotag/geotag_from_gopro.py
@@ -7,11 +7,11 @@ import typing as T
 from tqdm import tqdm
 
 from ..geo import get_max_distance_from_start, gps_distance, pairwise
-from .. import types, image_log, ffmpeg, error
+from .. import types, ffmpeg, error, utils
 from .geotag_from_gpx import GeotagFromGPXWithProgress
 from .geotag_from_generic import GeotagFromGeneric
 from .gpmf import parse_bin, interpolate_times
-from . import utils
+from . import utils as geotag_utils
 
 
 LOG = logging.getLogger(__name__)
@@ -27,7 +27,7 @@ class GeotagFromGoPro(GeotagFromGeneric):
     ):
         self.image_dir = image_dir
         if os.path.isdir(source_path):
-            self.videos = image_log.get_video_file_list(source_path, abs_path=True)
+            self.videos = utils.get_video_file_list(source_path, abs_path=True)
         else:
             # it is okay to not suffix with .mp4
             self.videos = [source_path]
@@ -38,11 +38,11 @@ class GeotagFromGoPro(GeotagFromGeneric):
     def to_description(self) -> T.List[types.ImageDescriptionFileOrError]:
         descs: T.List[types.ImageDescriptionFileOrError] = []
 
-        images = image_log.get_total_file_list(self.image_dir)
+        images = utils.get_total_file_list(self.image_dir)
         for video in self.videos:
             LOG.debug("Processing GoPro video: %s", video)
 
-            sample_images = image_log.filter_video_samples(images, video)
+            sample_images = utils.filter_video_samples(images, video)
             LOG.debug(
                 "Found %d sample images from video %s",
                 len(sample_images),
@@ -55,7 +55,7 @@ class GeotagFromGoPro(GeotagFromGeneric):
             points = get_points_from_gpmf(video)
 
             # bypass empty points to raise MapillaryGPXEmptyError
-            if points and utils.is_video_stationary(
+            if points and geotag_utils.is_video_stationary(
                 get_max_distance_from_start([(p.lat, p.lon) for p in points])
             ):
                 LOG.warning(
@@ -148,7 +148,7 @@ if __name__ == "__main__":
     import sys
 
     points = get_points_from_gpmf(sys.argv[1])
-    gpx = utils.convert_points_to_gpx(points)
+    gpx = geotag_utils.convert_points_to_gpx(points)
     print(gpx.to_xml())
 
     LOG.setLevel(logging.INFO)
@@ -157,7 +157,7 @@ if __name__ == "__main__":
     LOG.addHandler(handler)
     LOG.info(
         "Stationary: %s",
-        utils.is_video_stationary(
+        geotag_utils.is_video_stationary(
             get_max_distance_from_start([(p.lat, p.lon) for p in points])
         ),
     )
